backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
gunicorn
psycopg2-binary
python-dotenv
Werkzeug
SQLAlchemy